publishing/requirements.txt
GitPython==3.1.18
influxdb-client==1.27.0
